+++ js/image-scale.js
@@ -1,5 +1,3 @@
-// TODO: надо рефакторить
-
 const MIN_SIZE = 25;
 const MAX_SIZE = 100;
