+++ js/picture.js
@@ -1,8 +1,27 @@
 // TODO: надо рефакторить
-
-import {viewModal} from './modal.js';
-
 const socialComment = document.querySelector('.social__comment').cloneNode(true);
+
+const viewModal = () => {
+  const modal = document.querySelector('.big-picture');
+  const body = document.querySelector('body');
+  const cancel = document.querySelector('.big-picture__cancel');
+
+  modal.classList.remove('hidden');
+  body.classList.add('modal-open');
+
+  document.addEventListener('keydown', (evt) => {
+    if (evt.key === 'Escape') {
+      evt.preventDefault();
+      body.classList.remove('modal-open');
+      modal.classList.add('hidden');
+    }
+  });
+
+  cancel.addEventListener('click', () => {
+    body.classList.remove('modal-open');
+    modal.classList.add('hidden');
+  });
+};
 
 const viewPicture = (currentObject) => {
   viewModal();
