+++ js/slider-const.js
@@ -1,5 +1,3 @@
-// TODO: надо рефакторить
-
 const SLIDER_CONST_MAP = new Map([
   ['chrome',
     {
