+++ js/validate-functions.js
@@ -7,7 +7,7 @@ const splitString = (values) => {
 }
 
 const validateStartSimbol = (values) => {
-  const tags = splitString();
+  const tags = splitString(values);
 
   for(let i = 0; i < tags.length; i++) {
     if(tags[i] !== '' && tags[i][0] !== START_SIMBOL) {
@@ -19,7 +19,7 @@ const validateStartSimbol = (values) => {
 };
 
 const validateCorrectSimbol = (values) => {
-  const tags = splitString();
+  const tags = splitString(values);
 
   for(let i = 0; i < tags.length; i++) {
     if(tags[i] !== '' && !HASHTAG_REGEXP.test(tags[i])) {
@@ -31,13 +31,13 @@ const validateCorrectSimbol = (values) => {
 };
 
 const validateUniqueValue = (values) => {
-  const tags = splitString();
+  const tags = splitString(values);
 
   const set = new Set(tags);
 
   return set.size === tags.length;
 };
 
-const validateMaxCountValue = (values) => splitString().length <= MAX_COUNT_TAGS;
+const validateMaxCountValue = (values) => splitString(values).length <= MAX_COUNT_TAGS;
 
 export {validateStartSimbol, validateCorrectSimbol, validateUniqueValue, validateMaxCountValue};
